Added head container and it's children
The head container matches the site style and padding was added to blend
in but also space out the header from the container.

The container was made into a column so it will separate the information
and the image that will be added later.

.container <h1> need margin to be removed so that the right box and
left box align.

HTML- Added Left box and right box.